+++ internal/usecase/usecase_stream.go
@@ -2,13 +2,12 @@ package usecase
 
 import (
 	"context"
-	"fmt"
-	"sync"
 	"time"
 
 	"toc-machine-trading/internal/entity"
 	"toc-machine-trading/internal/usecase/rabbit"
 	"toc-machine-trading/internal/usecase/repo"
+	"toc-machine-trading/pkg/config"
 
 	"github.com/google/go-cmp/cmp"
 )
@@ -17,6 +16,10 @@ import (
 type StreamUseCase struct {
 	repo   StreamRepo
 	rabbit StreamRabbit
+
+	tradeSwitchCfg config.TradeSwitch
+	analyzeCfg     config.Analyze
+	basic          entity.BasicInfo
 }
 
 // NewStream -.
@@ -25,6 +28,15 @@ func NewStream(r *repo.StreamRepo, t *rabbit.StreamRabbit) {
 		repo:   r,
 		rabbit: t,
 	}
+
+	cfg, err := config.GetConfig()
+	if err != nil {
+		log.Panic(err)
+	}
+
+	uc.tradeSwitchCfg = cfg.TradeSwitch
+	uc.analyzeCfg = cfg.Analyze
+	uc.basic = *cc.GetBasicInfo()
 
 	go uc.ReceiveEvent(context.Background())
 	go uc.ReceiveOrderStatus(context.Background())
@@ -55,8 +67,14 @@ func (uc *StreamUseCase) ReceiveOrderStatus(ctx context.Context) {
 		for {
 			order := <-orderStatusChan
 			cacheOrder := cc.GetOrderByOrderID(order.OrderID)
+			order.TradeTime = cacheOrder.TradeTime
+
 			if !cmp.Equal(order, cacheOrder) {
 				cc.SetOrderByOrderID(order)
+			}
+
+			if err := uc.repo.InserOrUpdatetOrder(ctx, order); err != nil {
+				log.Error(err)
 			}
 		}
 	}()
@@ -67,89 +85,61 @@ func (uc *StreamUseCase) ReceiveOrderStatus(ctx context.Context) {
 func (uc *StreamUseCase) ReceiveStreamData(ctx context.Context, targetArr []*entity.Target) {
 	for _, t := range targetArr {
 		target := t
-		tickChan := make(chan *entity.RealTimeTick)
-		bidAskChan := make(chan *entity.RealTimeBidAsk)
+		data := &RealTimeData{
+			stockNum: target.StockNum,
+			orderMap: make(map[entity.OrderAction][]*entity.Order),
+			// quantity should decide by bisrate
+			orderQuantity: 1,
+			tickChan:      make(chan *entity.RealTimeTick),
+			bidAskChan:    make(chan *entity.RealTimeBidAsk),
+		}
 		finishChan := make(chan struct{})
-		go uc.tradeAgent(tickChan, bidAskChan, finishChan)
+		go uc.tradeAgent(data, finishChan)
 		for {
 			_, ok := <-finishChan
 			if !ok {
 				break
 			}
 		}
-		go uc.rabbit.TickConsumer(fmt.Sprintf("tick:%s", target.StockNum), tickChan)
-		go uc.rabbit.BidAskConsumer(fmt.Sprintf("bid_ask:%s", target.StockNum), bidAskChan)
+		go uc.rabbit.TickConsumer(target.StockNum, data.tickChan)
+		go uc.rabbit.BidAskConsumer(target.StockNum, data.bidAskChan)
 	}
 	bus.PublishTopicEvent(topicSubscribeTickTargets, ctx, targetArr)
 }
 
-func (uc *StreamUseCase) tradeAgent(tickChan chan *entity.RealTimeTick, bidAskChan chan *entity.RealTimeBidAsk, finishChan chan struct{}) {
-	var bidAsk *entity.RealTimeBidAsk
-	var currentStatus orderMap
+func (uc *StreamUseCase) tradeAgent(data *RealTimeData, finishChan chan struct{}) {
 	go func() {
 		for {
-			tick := <-tickChan
-			action := currentStatus.checkNeededPost()
-			switch action {
-			case entity.ActionNone:
-				if bidAsk != nil && bidAsk.AskPrice1 == tick.Close {
-					order := &entity.Order{
-						StockNum:  tick.StockNum,
-						Action:    entity.ActionBuy,
-						Price:     tick.Close,
-						Quantity:  1,
-						TradeTime: time.Now(),
-					}
-					bus.PublishTopicEvent(topicOrder, order)
-					go currentStatus.checkOrderStatus(order)
-				}
-			case entity.ActionWait:
+			tick := <-data.tickChan
+			data.tickArr = append(data.tickArr, tick)
+			order := data.generateOrder(uc.analyzeCfg)
+			if order == nil {
 				continue
 			}
+
+			bus.PublishTopicEvent(topicPlaceOrder, order)
+			data.waitingOrder = order
+
+			var timeout time.Duration
+			switch order.Action {
+			case entity.ActionBuy, entity.ActionSellFirst:
+				if time.Now().After(uc.basic.TradeDay.Add(time.Duration(uc.tradeSwitchCfg.TradeInEndTime) * time.Hour)) {
+					continue
+				}
+
+				timeout = time.Duration(uc.tradeSwitchCfg.TradeInEndTime) * time.Second
+			case entity.ActionSell, entity.ActionBuyLater:
+				timeout = time.Duration(uc.tradeSwitchCfg.TradeOutEndTime) * time.Second
+			}
+			go data.checkOrderStatus(order, timeout)
 		}
 	}()
 	go func() {
 		for {
-			bidAsk = <-bidAskChan
+			data.bidAsk = <-data.bidAskChan
 		}
 	}()
 
 	// close channel to start receive data from rabbitmq
 	close(finishChan)
-}
-
-type orderMap struct {
-	data     map[entity.OrderAction][]*entity.Order
-	mu       sync.RWMutex
-	checking bool
-}
-
-func (o *orderMap) checkOrderStatus(order *entity.Order) {
-	o.checking = true
-	for {
-		if order.OrderID != "" {
-			break
-		}
-		time.Sleep(time.Second)
-	}
-	o.mu.Lock()
-	o.data[order.Action] = append(o.data[order.Action], order)
-	o.mu.Unlock()
-	o.checking = false
-}
-
-func (o *orderMap) checkNeededPost() entity.OrderAction {
-	if o.checking {
-		return entity.ActionWait
-	}
-	o.mu.RLock()
-	defer o.mu.RUnlock()
-	if len(o.data[entity.ActionBuy]) > len(o.data[entity.ActionSell]) {
-		return entity.ActionSell
-	}
-
-	if len(o.data[entity.ActionSellFirst]) > len(o.data[entity.ActionBuyLater]) {
-		return entity.ActionBuyLater
-	}
-	return entity.ActionNone
 }